[Refactor] middleWare의 알림을 테이블 뷰 management가 직접 수신 (#27)
데이터를 조작하는 것은 middleWare가 직접 소통하는 다른 클래스가 하고,
데이터가 변경될 때마다 변경 사항을 NotificationCenter.default.post 로 전달합니다.
전달된 Notification은 테이블 뷰 혹은 뷰 컨트롤러가 아닌 management가 수신.
management는 UITableViewDelegate/DataSource 역할을 하므로
뷰에 화면 업데이트 지시할 수 있습니다.
