+++ Client/iOS/TodoList/ViewController.swift
@@ -12,7 +12,7 @@ class ViewController: UIViewController {
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-        AppDelegate.middleware.fetchAllCards()
+        AppDelegate.middleWare.fetchAllCards()
     }
 }
